+++ src/pages/dapp/components/RequestConfirmation/RequestConfirmation.tsx
@@ -24,6 +24,7 @@ import { walletAllAssetsState } from '../../../../recoil/atom';
 import { useLedgerStatus } from '../../../../hooks/useLedgerStatus';
 import { LEDGER_WALLET_TYPE } from '../../../../service/LedgerService';
 import { ledgerNotification } from '../../../../components/LedgerNotification/LedgerNotification';
+import GasStepSelectEVMDApp from '../../../../components/GasStepSelect/GasStepSelectEVMDApp';
 
 const { Content, Footer } = Layout;
 
@@ -35,7 +36,7 @@ interface RequestConfirmationProps {
   wallet: Wallet;
   visible: boolean;
   dapp?: Dapp;
-  onConfirm: () => void;
+  onConfirm: (info: { gasPrice: BigNumber, gasLimit: BigNumber }) => void;
   onCancel: () => void;
 }
 
@@ -58,6 +59,8 @@ const RequestConfirmation = (props: RequestConfirmationProps) => {
   const [isContractAddressReview, setIsContractAddressReview] = useState(false);
   const [isConfirmDisabled, setIsConfirmDisabled] = useState(false);
   const [allAssets, setAllAssets] = useRecoilState(walletAllAssetsState);
+  const [gasPrice, setGasPrice] = useState(new BigNumber((event.object as any)?.gasPrice ?? 0));
+  const [gasLimit, setGasLimit] = useState(new BigNumber((event.object as any)?.gas ?? 0));
 
   const { isLedgerConnected } = useLedgerStatus({ asset: cronosAsset });
 
@@ -83,7 +86,7 @@ const RequestConfirmation = (props: RequestConfirmationProps) => {
     setIsConfirmDisabled(false);
 
     if (event.name === 'signTransaction') {
-      const networkFee = event ? new BigNumber(event.object?.gas).times(event.object?.gasPrice) : 0;
+      const networkFee = event ? gasLimit.times(gasPrice) : 0;
       const total = event ? new BigNumber(event.object?.value ?? '0').plus(networkFee) : 0;
 
       const isDisabled = new BigNumber(cronosAsset?.balance ?? '0').isLessThan(total);
@@ -91,12 +94,10 @@ const RequestConfirmation = (props: RequestConfirmationProps) => {
 
       return (
         <>
-          <div className="row">
-            <div className="title">{t('dapp.requestConfirmation.networkFee.title')}</div>
-            <div>{`${scaledAmount(networkFee.toString(), cronosAsset?.decimals ?? 1)} ${
-              cronosAsset?.symbol
-            }`}</div>
-          </div>
+          <GasStepSelectEVMDApp asset={cronosAsset} gasLimit={gasLimit} gasPrice={gasPrice} onChange={(_gasLimit, _gasPrice) => {
+            setGasLimit(_gasLimit)
+            setGasPrice(_gasPrice)
+          }} />
           <div className="row">
             <div className="title">{t('dapp.requestConfirmation.total.title')}</div>
             <div>{`${scaledAmount(total.toString(), cronosAsset?.decimals ?? 1)} ${
@@ -149,9 +150,8 @@ const RequestConfirmation = (props: RequestConfirmationProps) => {
     }
 
     if (event.name === 'tokenApproval') {
-      const fee = new BigNumber(event.object.gas).times(event.object.gasPrice).toString();
-      const networkFee = fee;
-      const total = fee;
+      const networkFee = event ? gasLimit.times(gasPrice) : 0;
+      const total = networkFee;
       const { contractAddress } = event.object.tokenData;
 
       const isDisabled = new BigNumber(cronosAsset?.balance ?? '0').isLessThan(total);
@@ -159,12 +159,10 @@ const RequestConfirmation = (props: RequestConfirmationProps) => {
 
       return (
         <>
-          <div className="row">
-            <div className="title">{t('dapp.requestConfirmation.networkFee.title')}</div>
-            <div>{`${scaledAmount(networkFee.toString(), cronosAsset?.decimals ?? 1)} ${
-              cronosAsset?.symbol
-            }`}</div>
-          </div>
+          <GasStepSelectEVMDApp asset={cronosAsset} gasLimit={gasLimit} gasPrice={gasPrice} onChange={(_gasLimit, _gasPrice) => {
+            setGasLimit(_gasLimit)
+            setGasPrice(_gasPrice)
+          }} />
           <div className="row">
             <div className="title">{t('dapp.requestConfirmation.total.title')}</div>
             <div>{`${scaledAmount(total.toString(), cronosAsset?.decimals ?? 1)} ${
@@ -277,7 +275,12 @@ const RequestConfirmation = (props: RequestConfirmationProps) => {
             <Button
               type="primary"
               htmlType="submit"
-              onClick={onConfirm}
+              onClick={() => {
+                onConfirm({
+                  gasLimit,
+                  gasPrice
+                })
+              }}
               disabled={
                 isConfirmDisabled ||
                 (!isLedgerConnected && currentSession.wallet.walletType === LEDGER_WALLET_TYPE)