+++ src/service/TransactionRequestModels.ts
@@ -70,7 +70,7 @@ export interface DelegationRequest {
   walletType: string; // normal, ledger
 }
 
-export interface UndelegationRequest extends DelegationRequest {}
+export interface UndelegationRequest extends DelegationRequest { }
 export interface RedelegationRequest {
   validatorSourceAddress: string;
   validatorDestinationAddress: string;