+++ src/service/TransactionSenderService.ts
@@ -32,6 +32,7 @@ import {
 } from './TransactionRequestModels';
 import { StorageService } from '../storage/StorageService';
 import { CronosClient } from './cronos/CronosClient';
+import { EVMClient } from './rpc/clients/EVMClient';
 import { TransactionPrepareService } from './TransactionPrepareService';
 import { evmTransactionSigner } from './signers/EvmTransactionSigner';
 import { LEDGER_WALLET_TYPE, createLedgerDevice } from './LedgerService';
@@ -39,7 +40,7 @@ import { TransactionHistoryService } from './TransactionHistoryService';
 import { getCronosEvmAsset, sleep } from '../utils/utils';
 import { BridgeService } from './bridge/BridgeService';
 import { walletService } from './WalletService';
-import { EVMClient } from './rpc/clients/EVMClient';
+import { getCronosTendermintFeeConfig } from './Gas';
 
 export class TransactionSenderService {
   public readonly storageService: StorageService;
@@ -146,7 +147,7 @@ export class TransactionSenderService {
         } catch (e) {
           // eslint-disable-next-line no-console
           console.log(`ERROR_TRANSFERRING - ${currentAsset.assetType}`, e);
-          throw TypeError(e);
+          throw TypeError(e as any);
         }
 
       case UserAssetType.CRC_20_TOKEN:
@@ -247,12 +248,15 @@ export class TransactionSenderService {
             `ERROR_TRANSFERRING_TOKEN - ${currentAsset.assetType} ${currentAsset.symbol}`,
             e,
           );
-          throw TypeError(e);
+          throw TypeError(e as any);
         }
 
       case UserAssetType.TENDERMINT:
       case UserAssetType.IBC:
       case undefined: {
+
+        const { networkFee, gasLimit } = await getCronosTendermintFeeConfig()
+
         const {
           nodeRpc,
           accountNumber,
@@ -268,7 +272,7 @@ export class TransactionSenderService {
           memo: transferRequest.memo,
           accountNumber,
           accountSequence,
-          asset: currentAsset,
+          asset: transferRequest.asset,
         };
 
         let signedTxHex: string = '';
@@ -277,11 +281,15 @@ export class TransactionSenderService {
           signedTxHex = await ledgerTransactionSigner.signTransfer(
             transfer,
             transferRequest.decryptedPhrase,
+            networkFee,
+            gasLimit
           );
         } else {
           signedTxHex = await transactionSigner.signTransfer(
             transfer,
             transferRequest.decryptedPhrase,
+            networkFee,
+            gasLimit
           );
         }
 
@@ -332,15 +340,21 @@ export class TransactionSenderService {
     };
 
     let signedTxHex: string;
+    const { networkFee, gasLimit } = await getCronosTendermintFeeConfig()
+
     if (delegationRequest.walletType === LEDGER_WALLET_TYPE) {
       signedTxHex = await ledgerTransactionSigner.signDelegateTx(
         delegateTransaction,
         delegationRequest.decryptedPhrase,
+        networkFee,
+        gasLimit
       );
     } else {
       signedTxHex = await transactionSigner.signDelegateTx(
         delegateTransaction,
         delegationRequest.decryptedPhrase,
+        networkFee,
+        gasLimit
       );
     }
 
@@ -385,15 +399,21 @@ export class TransactionSenderService {
     };
 
     let signedTxHex: string;
+    const { networkFee, gasLimit } = await getCronosTendermintFeeConfig()
+
     if (undelegationRequest.walletType === LEDGER_WALLET_TYPE) {
       signedTxHex = await ledgerTransactionSigner.signUndelegateTx(
         undelegateTransaction,
         undelegationRequest.decryptedPhrase,
+        networkFee,
+        gasLimit
       );
     } else {
       signedTxHex = await transactionSigner.signUndelegateTx(
         undelegateTransaction,
         undelegationRequest.decryptedPhrase,
+        networkFee,
+        gasLimit
       );
     }
 
@@ -439,15 +459,21 @@ export class TransactionSenderService {
     };
 
     let signedTxHex: string;
+    const { networkFee, gasLimit } = await getCronosTendermintFeeConfig()
+
     if (redelegationRequest.walletType === LEDGER_WALLET_TYPE) {
       signedTxHex = await ledgerTransactionSigner.signRedelegateTx(
         redelegateTransactionUnsigned,
         redelegationRequest.decryptedPhrase,
+        networkFee,
+        gasLimit
       );
     } else {
       signedTxHex = await transactionSigner.signRedelegateTx(
         redelegateTransactionUnsigned,
         redelegationRequest.decryptedPhrase,
+        networkFee,
+        gasLimit
       );
     }
 
@@ -479,16 +505,21 @@ export class TransactionSenderService {
     };
 
     let signedTxHex: string = '';
+    const { networkFee, gasLimit } = await getCronosTendermintFeeConfig()
 
     if (voteRequest.walletType === LEDGER_WALLET_TYPE) {
       signedTxHex = await ledgerTransactionSigner.signVoteTransaction(
         voteTransactionUnsigned,
         voteRequest.decryptedPhrase,
+        networkFee,
+        gasLimit
       );
     } else {
       signedTxHex = await transactionSigner.signVoteTransaction(
         voteTransactionUnsigned,
         voteRequest.decryptedPhrase,
+        networkFee,
+        gasLimit
       );
     }
 
@@ -523,19 +554,25 @@ export class TransactionSenderService {
           accountSequence,
         };
 
-        let signedTxHex: string = '';
+        let signedTxHex = '';
+        const { networkFee, gasLimit } = await getCronosTendermintFeeConfig()
 
         if (nftTransferRequest.walletType === LEDGER_WALLET_TYPE) {
           signedTxHex = await ledgerTransactionSigner.signNFTTransfer(
             nftTransferUnsigned,
             nftTransferRequest.decryptedPhrase,
+            networkFee,
+            gasLimit
           );
         } else {
           signedTxHex = await transactionSigner.signNFTTransfer(
             nftTransferUnsigned,
             nftTransferRequest.decryptedPhrase,
+            networkFee,
+            gasLimit
           );
         }
+
         // It takes a few seconds for the indexing service to sync latest NFT state
         await sleep(7_000);
         await Promise.all([
@@ -657,15 +694,21 @@ export class TransactionSenderService {
 
     let signedTxHex: string;
 
+    const { gasLimit, networkFee } = await getCronosTendermintFeeConfig()
+
     if (rewardWithdrawRequest.walletType === LEDGER_WALLET_TYPE) {
       signedTxHex = await ledgerTransactionSigner.signWithdrawStakingRewardTx(
         withdrawStakingReward,
         rewardWithdrawRequest.decryptedPhrase,
+        networkFee,
+        gasLimit
       );
     } else {
       signedTxHex = await transactionSigner.signWithdrawStakingRewardTx(
         withdrawStakingReward,
         rewardWithdrawRequest.decryptedPhrase,
+        networkFee,
+        gasLimit
       );
     }
 
@@ -725,16 +768,22 @@ export class TransactionSenderService {
     };
 
     let signedTxHex: string = '';
+    const { networkFee, gasLimit } = await getCronosTendermintFeeConfig()
+
 
     if (nftMintRequest.walletType === LEDGER_WALLET_TYPE) {
       signedTxHex = await ledgerTransactionSigner.signNFTMint(
         nftMintUnsigned,
         nftMintRequest.decryptedPhrase,
+        networkFee,
+        gasLimit
       );
     } else {
       signedTxHex = await transactionSigner.signNFTMint(
         nftMintUnsigned,
         nftMintRequest.decryptedPhrase,
+        networkFee,
+        gasLimit
       );
     }
 
@@ -771,16 +820,21 @@ export class TransactionSenderService {
     };
 
     let signedTxHex: string = '';
+    const { networkFee, gasLimit } = await getCronosTendermintFeeConfig()
 
     if (nftDenomIssueRequest.walletType === LEDGER_WALLET_TYPE) {
       signedTxHex = await ledgerTransactionSigner.signNFTDenomIssue(
         nftDenomIssueUnsigned,
         nftDenomIssueRequest.decryptedPhrase,
+        networkFee,
+        gasLimit,
       );
     } else {
       signedTxHex = await transactionSigner.signNFTDenomIssue(
         nftDenomIssueUnsigned,
         nftDenomIssueRequest.decryptedPhrase,
+        networkFee,
+        gasLimit
       );
     }