+++ src/service/signers/SignerProvider.ts
@@ -1,9 +1,11 @@
 import { Bytes } from '@crypto-org-chain/chain-jslib/lib/dist/utils/bytes/bytes';
+import { DerivationPathStandard } from './LedgerSigner';
 
 export interface ISignerProvider {
   // cosmos
-  getPubKey(index: number, showLedgerDisplay: boolean): Promise<Bytes>;
-  getAddress(index: number, addressPrefix: string, showLedgerDisplay: boolean): Promise<string>;
+  getPubKey(index: number, derivationPathStandard: DerivationPathStandard, showLedgerDisplay: boolean): Promise<Bytes>;
+  getAddress(index: number, addressPrefix: string, derivationPathStandard: DerivationPathStandard, showLedgerDisplay: boolean): Promise<string>;
+  getAddressList(startIndex:number, gap: number, addressPrefix: string, derivationPathStandard: DerivationPathStandard): Promise<string[]>;
   sign(message: Bytes): Promise<Bytes>;
 
   // eth
@@ -17,7 +19,8 @@ export interface ISignerProvider {
     value: string,
     data: string,
   ): Promise<string>;
-  getEthAddress(index: number, display: boolean): Promise<string>;
+  getEthAddress(index: number, standard: DerivationPathStandard, display: boolean): Promise<string>;
+  getEthAddressList(startIndex:number, gap: number, standard: DerivationPathStandard): Promise<string[]>;
   signPersonalMessage(index: number, message: string): Promise<string>;
   signTypedDataV4(index: number, typedData: string): Promise<string>;
 }